GetTimeCount()を廃止
時刻とチックカウントは似て非なるもの

+++ BonCtrl/DropCount.h
@@ -49,7 +49,7 @@ protected:
 	ULONGLONG drop;
 	ULONGLONG scramble;
 	vector<wstring> log;
-	LONGLONG lastLogTime;
+	DWORD lastLogTime;
 	ULONGLONG lastLogDrop;
 	ULONGLONG lastLogScramble;
 	float signalLv;
